Moving levenshtein suggestion code around
