+++ src/parse/BUILD
@@ -15,8 +15,8 @@ cgo_library(
     deps = [
         ':builtin_rules',
         '//src/core',
+        '//src/utils',
         '//third_party/go:gcfg',
-        '//third_party/go:levenshtein',
         '//third_party/go:logging',
         '//third_party/go:osext',
     ],
